Extract one helper for getting dark theme
