+++ androidshared/src/main/java/org/odk/collect/androidshared/system/ContextExt.kt
@@ -1,10 +1,12 @@
 package org.odk.collect.androidshared.system
 
+import android.app.Activity
 import android.content.Context
+import android.content.res.Configuration
 import android.util.TypedValue
 import androidx.annotation.AttrRes
 
-object ContextUtils {
+object ContextExt {
 
     /**
      * Be careful when using this method to retrieve colors, especially for those defined
@@ -16,5 +18,11 @@ object ContextUtils {
         val outValue = TypedValue()
         context.theme.resolveAttribute(resId, outValue, true)
         return outValue.data
+    }
+
+    @JvmStatic
+    fun Context.isDarkTheme(): Boolean {
+        val uiMode: Int = this.resources.configuration.uiMode
+        return (uiMode and Configuration.UI_MODE_NIGHT_MASK) == Configuration.UI_MODE_NIGHT_YES
     }
 }
