+++ collect_app/src/main/java/org/odk/collect/android/views/DayNightProgressDialog.java
@@ -1,5 +1,7 @@
 package org.odk.collect.android.views;
 
+import static org.odk.collect.androidshared.system.ContextExt.isDarkTheme;
+
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -21,7 +23,7 @@ import org.odk.collect.material.MaterialProgressDialogFragment;
 public class DayNightProgressDialog extends ProgressDialog {
 
     public DayNightProgressDialog(Context context) {
-        super(context, new ThemeUtils(context).isDarkTheme() ? android.R.style.Theme_Material_Dialog_Alert : android.R.style.Theme_Material_Light_Dialog_Alert);
+        super(context, isDarkTheme(context) ? android.R.style.Theme_Material_Dialog_Alert : android.R.style.Theme_Material_Light_Dialog_Alert);
 
         setOnShowListener(dialog -> {
             ((ProgressDialog) dialog).getButton(DialogInterface.BUTTON_POSITIVE).setTextColor(new ThemeUtils(((ProgressDialog) dialog).getContext()).getColorPrimary());
